Move some files to obstacle dir
