+++ consai_robot_controller/CMakeLists.txt
@@ -39,7 +39,7 @@ add_library(controller_component SHARED
   src/constraint_parser.cpp
   src/tactic_control_ball.cpp
   src/tactic_obstacle_avoidance.cpp
-  src/obstacle_observer.cpp
+  src/obstacle/obstacle_observer.cpp
 )
 target_compile_definitions(controller_component
   PRIVATE "CONSAI_ROBOT_CONTROLLER_BUILDING_DLL")
